all archives now compiling
had to make a few changes to rapidjson since it uses constexprs in one location,
also had to make sure we used the correct sprintf in a place where we inserted the code

+++ include/cereal/archives/json.hpp
@@ -325,7 +325,7 @@ namespace cereal
       void loadValue(bool & val)        { val = itsValueStack.back().value().GetBool();   ++itsValueStack.back(); }
       void loadValue(int64_t & val)     { val = itsValueStack.back().value().GetInt64();  ++itsValueStack.back(); }
       void loadValue(uint64_t & val)    { val = itsValueStack.back().value().GetUint64(); ++itsValueStack.back(); }
-      void loadValue(float & val)       { val = itsValueStack.back().value().GetDouble(); ++itsValueStack.back(); }
+      void loadValue(float & val)       { val = static_cast<float>(itsValueStack.back().value().GetDouble()); ++itsValueStack.back(); }
       void loadValue(double & val)      { val = itsValueStack.back().value().GetDouble(); ++itsValueStack.back(); }
       void loadValue(std::string & val) { val = itsValueStack.back().value().GetString(); ++itsValueStack.back(); }
